move inspector-controls folder into edit folder
